reduce duplication,pushing methods to parent class
Currency variable is added to Money class and getter 'currency()' is implemented for it
So it helps to avoid code duplication in subclasses.
Currency is now initialized in factory methods, during objects initialization.
'times()' method is also using now Money factory methods, to return a new Dollar of Franc object